Change SalaryPerDay type from decimal to int

+++ Models/Employee.cs
@@ -5,6 +5,6 @@ namespace api_demo.Models
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Department { get; set; } = string.Empty;
-        public decimal SalaryPerDay{ get; set; }  
+        public int SalaryPerDay{ get; set; }  
     }
 }
